Improved auto-updater
Also fixed a minor bug that happened on some platforms which cause the auto-updater script to fail. This was due to run.sh being overwritten while running. 

Fixed this by making run.sh not update itself. Which shouldn't be an issue since the run.sh script should not need to be updated again.